+++ packages/perseus/src/widgets/interactive-graphs/axis-ticks.tsx
@@ -120,12 +120,8 @@ export const AxisTicks = () => {
         height,
     };
 
-    // TODO(benchristel): destructure these in one line
-    const [xMin, xMax] = range[X];
-    const [yMin, yMax] = range[Y];
-
-    const yTickStep = tickStep[Y];
-    const xTickStep = tickStep[X];
+    const [[xMin, xMax], [yMin, yMax]] = range;
+    const [xTickStep, yTickStep] = tickStep;
 
     const yGridTicks = generateTickLocations(yTickStep, yMin, yMax);
     const xGridTicks = generateTickLocations(xTickStep, xMin, xMax);